Move SecretSharing to a separate class
